workshop_booking-master/requirements.txt
Django==3.0.7
coverage
pyaml
django-recurrence
python-decouple==3.3
pandas
djangorestframework==3.14.0
django-cors-headers
